src/calculator/requirements.txt
# AI FinOps Cost Calculator Dependencies
flask>=3.0.0
prometheus-client>=0.20.0
requests>=2.31.0
pyyaml>=6.0.1

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.4.0